requirements.txt
flask
ultralytics
opencv-python
numpy
pillow
keras
mediapipe
scipy
tensorflow
